parsing/requirements.txt
datefinder==0.6.0
dateparser==0.4.0
egenix-mx-base==3.2.9
jdatetime==1.8.1
nltk==3.2.1
python-dateutil==2.5.3
pytz==2016.6.1
regex==2016.7.21
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.14
six==1.10.0
umalqurra==0.2
